+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@
 - `[Lookup]` Resynced the lookup api settings, methods and events with the core component. Also made sure everything is using ngZone. `TJM` ([Issue #236](https://github.com/infor-design/enterprise/issues/236))
 - `[Lookup]` And added an example showing how to use the buttonsetAPI to enable and disable buttons.  `TJM` ([Issue #236](https://github.com/infor-design/enterprise/issues/236))
 - `[Lookup]` And added an example showing how to use a dataset as an input to populate the lookup.  `TJM` ([Issue #236](https://github.com/infor-design/enterprise/issues/236))
+- `[Datagrid]` Build expanded datagrid row with dynamic component. Special case: nested datagrid.  `TJM` ([Issue #206](https://github.com/infor-design/enterprise-ng/issues/206))
 
 ## v7.2.1